Move cards right now fully functional
Annoying error of the card slots being in wrong position in the list so moving cards one to the right would act up. now list of lists uses nulls instead of removing the cards, and then nulls are swapped to left as to only move cards by one to the right.